# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Paquito Coach API",
    "version": "0.0.0-dev",
    "description": "API REST del MVP. Autenticación por cookie de sesión."
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/exercises": {
      "get": {
        "tags": [
          "entrenamiento"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "chest",
                "lats",
                "middleBack",
                "lowerBack",
                "traps",
                "shoulders",
                "biceps",
                "triceps",
                "forearms",
                "abs",
                "quads",
                "hamstrings",
                "glutes",
                "calves",
                "adductors",
                "abductors",
                "neck",
                "other"
              ]
            },
            "in": "query",
            "name": "muscle",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "barbell",
                "dumbbell",
                "machine",
                "cable",
                "bodyweight",
                "kettlebell",
                "bands",
                "other"
              ]
            },
            "in": "query",
            "name": "equipment",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            },
            "in": "query",
            "name": "own",
            "required": false
          },
          {
            "schema": {
              "default": 40,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "aliases": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "muscle": {
                        "type": "string",
                        "enum": [
                          "chest",
                          "lats",
                          "middleBack",
                          "lowerBack",
                          "traps",
                          "shoulders",
                          "biceps",
                          "triceps",
                          "forearms",
                          "abs",
                          "quads",
                          "hamstrings",
                          "glutes",
                          "calves",
                          "adductors",
                          "abductors",
                          "neck",
                          "other"
                        ]
                      },
                      "secondary": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "enum": [
                            "chest",
                            "lats",
                            "middleBack",
                            "lowerBack",
                            "traps",
                            "shoulders",
                            "biceps",
                            "triceps",
                            "forearms",
                            "abs",
                            "quads",
                            "hamstrings",
                            "glutes",
                            "calves",
                            "adductors",
                            "abductors",
                            "neck",
                            "other"
                          ]
                        }
                      },
                      "equipment": {
                        "type": "string",
                        "enum": [
                          "barbell",
                          "dumbbell",
                          "machine",
                          "cable",
                          "bodyweight",
                          "kettlebell",
                          "bands",
                          "other"
                        ]
                      },
                      "instructions": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "videoUrl": {
                        "nullable": true,
                        "type": "string"
                      },
                      "own": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "aliases",
                      "muscle",
                      "secondary",
                      "equipment",
                      "instructions",
                      "videoUrl",
                      "own"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "entrenamiento"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "muscle": {
                    "type": "string",
                    "enum": [
                      "chest",
                      "lats",
                      "middleBack",
                      "lowerBack",
                      "traps",
                      "shoulders",
                      "biceps",
                      "triceps",
                      "forearms",
                      "abs",
                      "quads",
                      "hamstrings",
                      "glutes",
                      "calves",
                      "adductors",
                      "abductors",
                      "neck",
                      "other"
                    ]
                  },
                  "equipment": {
                    "type": "string",
                    "enum": [
                      "barbell",
                      "dumbbell",
                      "machine",
                      "cable",
                      "bodyweight",
                      "kettlebell",
                      "bands",
                      "other"
                    ]
                  },
                  "instructions": {
                    "default": [],
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 500
                    }
                  },
                  "videoUrl": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 300,
                    "format": "uri"
                  }
                },
                "required": [
                  "name",
                  "muscle",
                  "equipment"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "aliases": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "muscle": {
                      "type": "string",
                      "enum": [
                        "chest",
                        "lats",
                        "middleBack",
                        "lowerBack",
                        "traps",
                        "shoulders",
                        "biceps",
                        "triceps",
                        "forearms",
                        "abs",
                        "quads",
                        "hamstrings",
                        "glutes",
                        "calves",
                        "adductors",
                        "abductors",
                        "neck",
                        "other"
                      ]
                    },
                    "secondary": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "chest",
                          "lats",
                          "middleBack",
                          "lowerBack",
                          "traps",
                          "shoulders",
                          "biceps",
                          "triceps",
                          "forearms",
                          "abs",
                          "quads",
                          "hamstrings",
                          "glutes",
                          "calves",
                          "adductors",
                          "abductors",
                          "neck",
                          "other"
                        ]
                      }
                    },
                    "equipment": {
                      "type": "string",
                      "enum": [
                        "barbell",
                        "dumbbell",
                        "machine",
                        "cable",
                        "bodyweight",
                        "kettlebell",
                        "bands",
                        "other"
                      ]
                    },
                    "instructions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "videoUrl": {
                      "nullable": true,
                      "type": "string"
                    },
                    "own": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "aliases",
                    "muscle",
                    "secondary",
                    "equipment",
                    "instructions",
                    "videoUrl",
                    "own"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/routines/{id}/assign": {
      "post": {
        "tags": [
          "entrenamiento"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientIds": {
                    "minItems": 1,
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "dates": {
                    "minItems": 1,
                    "maxItems": 120,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                    }
                  }
                },
                "required": [
                  "clientIds",
                  "dates"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "created"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}